new sql

+++ BE/services/IRoomService.cs
@@ -5,12 +5,13 @@ namespace BE.Services
     public interface IRoomService
     {
         Task<List<RoomDto>> GetAllAsync();
-        Task<RoomDto?> GetByIdAsync(long id);
+
+        Task<RoomDto?> GetByIdAsync(int id);
 
         Task<RoomDto> CreateAsync(CreateRoomDto dto);
 
-        Task<RoomDto?> UpdateAsync(long id, UpdateRoomDto dto);
+        Task<RoomDto?> UpdateAsync(int id, UpdateRoomDto dto);
 
-        Task<bool> DeleteAsync(long id);
+        Task<bool> DeleteAsync(int id);
     }
 }
